feat(tickets): add the fetch function to the context file
